# Conflicts:
#	_maps/map_files/generic/z2.dmm
#	icons/obj/closet.dmi

+++ code/game/machinery/camera/camera.dm
@@ -10,7 +10,7 @@
 	use_power = 2
 	idle_power_usage = 5
 	active_power_usage = 10
-	layer = 5
+	layer = WALL_OBJ_LAYER
 
 	var/health = 50
 	var/list/network = list("SS13")
@@ -28,7 +28,6 @@
 	var/view_range = 7
 	var/short_range = 2
 
-	var/light_disabled = 0
 	var/alarm_on = 0
 	var/busy = 0
 	var/emped = 0  //Number of consecutive EMP's on this camera
@@ -42,6 +41,7 @@
 	assembly.state = 4
 	cameranet.cameras += src
 	cameranet.addCamera(src)
+	add_to_proximity_list(src, 1) //1 was default of everything
 	/* // Use this to look for cameras that have the same c_tag.
 	for(var/obj/machinery/camera/C in cameranet.cameras)
 		var/list/tempnetwork = C.network&src.network
@@ -51,14 +51,19 @@
 
 /obj/machinery/camera/initialize()
 	if(z == 1 && prob(3) && !start_active)
-		deactivate()
+		toggle_cam()
+
+/obj/machinery/camera/Move()
+	remove_from_proximity_list(src, 1)
+	return ..()
 
 /obj/machinery/camera/Destroy()
-	deactivate(null, 0) //kick anyone viewing out
+	toggle_cam(null, 0) //kick anyone viewing out
+	remove_from_proximity_list(src, 1)
 	if(assembly)
 		qdel(assembly)
 		assembly = null
-	if(istype(bug))
+	if(bug)
 		bug.bugged_cameras -= src.c_tag
 		if(bug.current == src)
 			bug.current = null
@@ -69,33 +74,34 @@
 	return ..()
 
 /obj/machinery/camera/emp_act(severity)
+	if(!status)
+		return
 	if(!isEmpProof())
 		if(prob(150/severity))
-			icon_state = "[initial(icon_state)]emp"
+			update_icon()
 			var/list/previous_network = network
 			network = list()
 			cameranet.removeCamera(src)
 			stat |= EMPED
 			SetLuminosity(0)
 			emped = emped+1  //Increase the number of consecutive EMP's
+			update_icon()
 			var/thisemp = emped //Take note of which EMP this proc is for
 			spawn(900)
 				if(loc) //qdel limbo
 					triggerCameraAlarm() //camera alarm triggers even if multiple EMPs are in effect.
 					if(emped == thisemp) //Only fix it if the camera hasn't been EMP'd again
 						network = previous_network
-						icon_state = initial(icon_state)
 						stat &= ~EMPED
+						update_icon()
 						if(can_use())
 							cameranet.addCamera(src)
 						emped = 0 //Resets the consecutive EMP count
-						spawn(100)
-							if(!qdeleted(src))
-								cancelCameraAlarm()
+						addtimer(src, "cancelCameraAlarm", 100)
 			for(var/mob/O in mob_list)
 				if (O.client && O.client.eye == src)
 					O.unset_machine()
-					O.reset_view(null)
+					O.reset_perspective(null)
 					O << "The screen bursts into static."
 			..()
 
@@ -103,9 +109,13 @@
 /obj/machinery/camera/ex_act(severity, target)
 	if(src.invuln)
 		return
-	else
-		..()
-	return
+	switch(severity)
+		if(1)
+			qdel(src)
+		if(2)
+			take_damage(50, BRUTE, 0)
+		else
+			take_damage(rand(30,60), BRUTE, 0)
 
 /obj/machinery/camera/proc/setViewRange(num = 7)
 	src.view_range = num
@@ -115,17 +125,6 @@
 	if(!istype(user))
 		return
 	user.electrocute_act(10, src)
-
-/obj/machinery/camera/attack_paw(mob/living/carbon/alien/humanoid/user)
-	if(!istype(user))
-		return
-	user.do_attack_animation(src)
-	add_hiddenprint(user)
-	visible_message("<span class='warning'>\The [user] slashes at [src]!</span>")
-	playsound(src.loc, 'sound/weapons/slash.ogg', 100, 1)
-	health = max(0, health - 30)
-	if(!health && status)
-		deactivate(user, 0)
 
 /obj/machinery/camera/attackby(obj/W, mob/living/user, params)
 	var/msg = "<span class='notice'>You attach [W] into the assembly's inner circuits.</span>"
@@ -140,12 +139,14 @@
 
 	if(panel_open)
 		if(istype(W, /obj/item/weapon/wirecutters)) //enable/disable the camera
-			deactivate(user, 1)
+			toggle_cam(user, 1)
 			health = initial(health) //this is a pretty simplistic way to heal the camera, but there's no reason for this to be complex.
+			return
 
 		else if(istype(W, /obj/item/device/multitool)) //change focus
 			setViewRange((view_range == initial(view_range)) ? short_range : initial(view_range))
 			user << "<span class='notice'>You [(view_range == initial(view_range)) ? "restore" : "mess up"] the camera's focus.</span>"
+			return
 
 		else if(istype(W, /obj/item/weapon/weldingtool))
 			if(weld(W, user))
@@ -154,10 +155,10 @@
 					assembly = new()
 				assembly.loc = src.loc
 				assembly.state = 1
-				assembly.dir = src.dir
+				assembly.setDir(src.dir)
 				assembly = null
 				qdel(src)
-				return
+			return
 
 		else if(istype(W, /obj/item/device/analyzer))
 			if(!isXRay())
@@ -166,6 +167,7 @@
 				user << "[msg]"
 			else
 				user << "[msg2]"
+			return
 
 		else if(istype(W, /obj/item/stack/sheet/mineral/plasma))
 			if(!isEmpProof())
@@ -174,6 +176,8 @@
 				qdel(W)
 			else
 				user << "[msg2]"
+			return
+
 		else if(istype(W, /obj/item/device/assembly/prox_sensor))
 			if(!isMotion())
 				upgradeMotion()
@@ -181,6 +185,7 @@
 				qdel(W)
 			else
 				user << "[msg2]"
+			return
 
 	// OTHER
 	if((istype(W, /obj/item/weapon/paper) || istype(W, /obj/item/device/pda)) && isliving(user))
@@ -213,36 +218,58 @@
 			else if (O.client && O.client.eye == src)
 				O << "[U] holds \a [itemname] up to one of the cameras ..."
 				O << browse(text("<HTML><HEAD><TITLE>[]</TITLE></HEAD><BODY><TT>[]</TT></BODY></HTML>", itemname, info), text("window=[]", itemname))
+		return
 
-	else if (istype(W, /obj/item/device/camera_bug))
-		if (!src.can_use())
+	else if(istype(W, /obj/item/device/camera_bug))
+		if(!can_use())
 			user << "<span class='notice'>Camera non-functional.</span>"
 			return
-		if(istype(src.bug))
+		if(bug)
 			user << "<span class='notice'>Camera bug removed.</span>"
-			src.bug.bugged_cameras -= src.c_tag
-			src.bug = null
+			bug.bugged_cameras -= src.c_tag
+			bug = null
 		else
 			user << "<span class='notice'>Camera bugged.</span>"
-			src.bug = W
-			src.bug.bugged_cameras[src.c_tag] = src
+			bug = W
+			bug.bugged_cameras[src.c_tag] = src
+		return
 
-	else if(istype(W, /obj/item/device/laser_pointer))
-		var/obj/item/device/laser_pointer/L = W
-		L.laser_act(src, user)
+	else if(istype(W, /obj/item/weapon/pai_cable))
+		var/obj/item/weapon/pai_cable/cable = W
+		cable.plugin(src, user)
+		return
 
+	return ..()
+
+/obj/machinery/camera/take_damage(damage, damage_type = BRUTE, sound_effect = 1)
+	switch(damage_type)
+		if(BRUTE)
+			if(sound_effect)
+				if(damage)
+					playsound(src, 'sound/weapons/smash.ogg', 50, 1)
+				else
+					playsound(src, 'sound/weapons/tap.ogg', 50, 1)
+		if(BURN)
+			if(sound_effect)
+				playsound(src.loc, 'sound/items/Welder.ogg', 100, 1)
+		else
+			return
+	if(damage < 10) //camera has a damage resistance threshold
+		return
+	health = max(0, health - damage)
+	if(!health && status)
+		triggerCameraAlarm()
+		toggle_cam(null, 0)
+
+/obj/machinery/camera/update_icon()
+	if(!status)
+		icon_state = "[initial(icon_state)]1"
+	else if (stat & EMPED)
+		icon_state = "[initial(icon_state)]emp"
 	else
-		if(W.force >= 10) //fairly simplistic, but will do for now.
-			user.changeNext_move(CLICK_CD_MELEE)
-			visible_message("<span class='warning'>[user] hits [src] with [W]!</span>", "<span class='warning'>You hit [src] with [W]!</span>")
-			health = max(0, health - W.force)
-			user.do_attack_animation(src)
-			if(!health && status)
-				triggerCameraAlarm()
-				deactivate(user, 1)
-	return
+		icon_state = "[initial(icon_state)]"
 
-/obj/machinery/camera/proc/deactivate(mob/user, displaymessage = 1) //this should be called toggle() but doing a find and replace for this would be ass
+/obj/machinery/camera/proc/toggle_cam(mob/user, displaymessage = 1)
 	status = !status
 	if(can_use())
 		cameranet.addCamera(src)
@@ -251,15 +278,10 @@
 		cameranet.removeCamera(src)
 	cameranet.updateChunk(x, y, z)
 	var/change_msg = "deactivates"
-	if(!status)
-		icon_state = "[initial(icon_state)]1"
-	else
-		icon_state = initial(icon_state)
+	if(status)
 		change_msg = "reactivates"
 		triggerCameraAlarm()
-		spawn(100)
-			if(!qdeleted(src))
-				cancelCameraAlarm()
+		addtimer(src, "cancelCameraAlarm", 100)
 	if(displaymessage)
 		if(user)
 			visible_message("<span class='danger'>[user] [change_msg] [src]!</span>")
@@ -268,6 +290,7 @@
 			visible_message("<span class='danger'>\The [src] [change_msg]!</span>")
 
 		playsound(src.loc, 'sound/items/Wirecutter.ogg', 100, 1)
+	update_icon()
 
 	// now disconnect anyone using the camera
 	//Apparently, this will disconnect anyone even if the camera was re-activated.
@@ -275,7 +298,7 @@
 	for(var/mob/O in player_list)
 		if (O.client && O.client.eye == src)
 			O.unset_machine()
-			O.reset_view(null)
+			O.reset_perspective(null)
 			O << "The screen bursts into static."
 
 /obj/machinery/camera/proc/triggerCameraAlarm()
@@ -306,20 +329,20 @@
 
 /atom/proc/auto_turn()
 	//Automatically turns based on nearby walls.
-	var/turf/simulated/wall/T = null
+	var/turf/closed/wall/T = null
 	for(var/i = 1, i <= 8; i += i)
 		T = get_ranged_target_turf(src, i, 1)
 		if(istype(T))
 			//If someone knows a better way to do this, let me know. -Giacom
 			switch(i)
 				if(NORTH)
-					src.dir = SOUTH
+					src.setDir(SOUTH)
 				if(SOUTH)
-					src.dir = NORTH
+					src.setDir(NORTH)
 				if(WEST)
-					src.dir = EAST
+					src.setDir(EAST)
 				if(EAST)
-					src.dir = WEST
+					src.setDir(WEST)
 			break
 
 //Return a working camera that can see a given mob
@@ -356,12 +379,19 @@
 	busy = 0
 	return 0
 
-/obj/machinery/camera/bullet_act(obj/item/projectile/proj)
-	if(proj.damage_type == BRUTE)
-		health = max(0, health - proj.damage)
-		if(!health && status)
-			triggerCameraAlarm()
-			deactivate(null, 1)
+/obj/machinery/camera/proc/Togglelight(on=0)
+	for(var/mob/living/silicon/ai/A in ai_list)
+		for(var/obj/machinery/camera/cam in A.lit_cameras)
+			if(cam == src)
+				return
+	if(on)
+		src.SetLuminosity(AI_CAMERA_LUMINOSITY)
+	else
+		src.SetLuminosity(0)
+
+/obj/machinery/camera/bullet_act(obj/item/projectile/P)
+	. = ..()
+	take_damage(P.damage, P.damage_type, 0)
 
 /obj/machinery/camera/portable //Cameras which are placed inside of things, such as helmets.
 	var/turf/prev_turf
@@ -376,3 +406,17 @@
 	if(cameranet && get_turf(src) != prev_turf)
 		cameranet.updatePortableCamera(src)
 		prev_turf = get_turf(src)
+
+/obj/machinery/camera/get_remote_view_fullscreens(mob/user)
+	if(view_range == short_range) //unfocused
+		user.overlay_fullscreen("remote_view", /obj/screen/fullscreen/impaired, 2)
+
+/obj/machinery/camera/update_remote_sight(mob/living/user)
+	user.see_invisible = SEE_INVISIBLE_LIVING //can't see ghosts through cameras
+	if(isXRay())
+		user.sight |= (SEE_TURFS|SEE_MOBS|SEE_OBJS)
+		user.see_in_dark = max(user.see_in_dark, 8)
+	else
+		user.sight = 0
+		user.see_in_dark = 2
+	return 1
