+++ code/game/objects/structures/morgue.dm
@@ -44,8 +44,7 @@
 /obj/structure/bodycontainer/relaymove(mob/user)
 	if(user.stat || !isturf(loc))
 		return
-	if(!open())
-		open()
+	open()
 
 /obj/structure/bodycontainer/attack_paw(mob/user)
 	return src.attack_hand(user)
@@ -64,7 +63,8 @@
 	add_fingerprint(user)
 
 /obj/structure/bodycontainer/attackby(obj/P, mob/user, params)
-	if (istype(P, /obj/item/weapon/pen))
+	add_fingerprint(user)
+	if(istype(P, /obj/item/weapon/pen))
 		var/t = stripped_input(user, "What would you like the label to be?", text("[]", name), null)
 		if (user.get_active_hand() != P)
 			return
@@ -74,7 +74,8 @@
 			name = text("[]- '[]'", initial(name), t)
 		else
 			name = initial(name)
-	add_fingerprint(user)
+	else
+		return ..()
 
 /obj/structure/bodycontainer/container_resist()
 	open()
@@ -82,17 +83,20 @@
 /obj/structure/bodycontainer/proc/open()
 	playsound(src.loc, 'sound/items/Deconstruct.ogg', 50, 1)
 	var/turf/T = get_step(src, opendir)
-	for(var/atom/movable/A in src)
-		A.loc = T
+	for(var/atom/movable/AM in src)
+		AM.forceMove(T)
 	update_icon()
 
 /obj/structure/bodycontainer/proc/close()
 	playsound(src.loc, 'sound/items/Deconstruct.ogg', 50, 1)
-	for(var/atom/movable/A in connected.loc)
-		if(!A.anchored || A == connected)
-			A.loc = src
+	for(var/atom/movable/AM in connected.loc)
+		if(!AM.anchored || AM == connected)
+			AM.forceMove(src)
 	update_icon()
 
+/obj/structure/bodycontainer/get_remote_view_fullscreens(mob/user)
+	if(user.stat == DEAD || !(user.sight & (SEEOBJS|SEEMOBS)))
+		user.overlay_fullscreen("remote_view", /obj/screen/fullscreen/impaired, 2)
 /*
  * Morgue
  */
@@ -108,24 +112,20 @@
 	..()
 
 /obj/structure/bodycontainer/morgue/update_icon()
-	if (!connected || connected.loc != src) //open or the tray broke off somehow
-		src.icon_state = "morgue0"
+	if (!connected || connected.loc != src) // Open or tray is gone.
+		icon_state = "morgue0"
 	else
-		if(src.contents.len == 1) //empty except for the tray
-			src.icon_state = "morgue1"
+		if(contents.len == 1)  // Empty
+			icon_state = "morgue1"
 		else
-
-			src.icon_state = "morgue2"//default dead no-client mob
-
-			var/list/compiled = recursive_mob_check(src,0,0)//run through contents
-
-			if(!length(compiled))//no mobs at all, but objects inside
-				src.icon_state = "morgue3"
+			icon_state = "morgue2" // Dead, brainded mob.
+			var/list/compiled = recursive_mob_check(src, 0, 0) // Search for mobs in all contents.
+			if(!length(compiled)) // No mobs?
+				icon_state = "morgue3"
 				return
-
 			for(var/mob/living/M in compiled)
 				if(M.client)
-					src.icon_state = "morgue4"//clone that mofo
+					icon_state = "morgue4" // Cloneable
 					break
 
 /*
@@ -180,7 +180,7 @@ var/global/list/crematoriums = new/list()
 		update_icon()
 
 		for(var/mob/living/M in contents)
-			if (M.stat!=2)
+			if (M.stat != DEAD)
 				M.emote("scream")
 			if(user)
 				user.attack_log +="\[[time_stamp()]\] Cremated <b>[M]/[M.ckey]</b>"
@@ -188,8 +188,9 @@ var/global/list/crematoriums = new/list()
 			else
 				log_attack("\[[time_stamp()]\] <b>UNKNOWN</b> cremated <b>[M]/[M.ckey]</b>")
 			M.death(1)
-			M.ghostize()
-			qdel(M)
+			if(M) //some animals get automatically deleted on death.
+				M.ghostize()
+				qdel(M)
 
 		for(var/obj/O in contents) //obj instead of obj/item so that bodybags and ashes get destroyed. We dont want tons and tons of ash piling up
 			if(O != connected) //Creamtorium does not burn hot enough to destroy the tray
@@ -210,7 +211,7 @@ var/global/list/crematoriums = new/list()
 /obj/structure/tray
 	icon = 'icons/obj/stationobjs.dmi'
 	density = 1
-	layer = 2.9
+	layer = BELOW_OBJ_LAYER
 	var/obj/structure/bodycontainer/connected = null
 	anchored = 1
 	pass_flags = LETPASSTHROW
@@ -266,7 +267,8 @@ var/global/list/crematoriums = new/list()
 	icon_state = "morguet"
 
 /obj/structure/tray/m_tray/CanPass(atom/movable/mover, turf/target, height=0)
-	if(height==0) return 1
+	if(height == 0)
+		return 1
 
 	if(istype(mover) && mover.checkpass(PASSTABLE))
 		return 1
@@ -274,3 +276,9 @@ var/global/list/crematoriums = new/list()
 		return 1
 	else
 		return 0
+
+/obj/structure/tray/m_tray/CanAStarPass(ID, dir, caller)
+	. = !density
+	if(ismovableatom(caller))
+		var/atom/movable/mover = caller
+		. = . || mover.checkpass(PASSTABLE)
